+++ tutorial/01_aircraft/go/pkg/define/define.go
@@ -1,7 +1,7 @@
 package define
 
 var (
-	WinWidth        = float32(1080)
-	WinHeight       = float32(1920)
+	WinWidth        = float64(1080)
+	WinHeight       = float64(1920)
 	Score     int64 = 0
 )
